Switch to Tornado based server
This commit removes Flask and replaces it with Tornado. Tornado
is based on asyncio and event loop based methodology which makes
it much easier to deal with concurrent requests and allows it to
service many more requests simultaneously.

Additionally, this commit includes a switch to a proper
multiprocessor pool of workers to handle the image manipulation.
This allows us to queue requests and hopefully handle more of them
at the same time in a controlled fashion.

All of the functionality in this new version of the server could
have been implemented in Flask and uWSGI but it is made significantly
easier by using Tornado and the asyncio semantics.