Support region requests that don't originate at a number divisble by 16
Due to the magic way jpegtran works it can't crop if the x or y values
aren't divisible by 16. Therefore in these cases we need to use pillow.
To make the cropping as fast as possible we actually crop down to near
the required crop using jpegtran and then use pillow to finish the job.