Refactor to be generic
Remove vfactor and mam specific code and provide a type and source
system. The configuration file is now used to define the types and
how the source image is retrieved for each of those types. The
number of types is infinite and each type can have a single source
type. The available source types currently are disk, web and
trusted_web.