Switch from tornado to fastapi [first pass]
fastapi + uvicorn is nicer to work with than tornado and has feels a
little more flask like which lessens the learning curve for developers
who aren't as familiar with tornado. Plus it's really fast and their
docs are great!

Note that the code is a bit messy but this is a parity commit where
it works on fastapi but hasn't been tidied up.