Significant refactor to accommodate mss (homeland) requirements
This commit refactors how image metadata is handled but leaves the
handling of the actual image processing largely untouched.

The method of image data and metadata retrieval is now handled in
the profiles module where the disk type persists from previous
versions of this code but the new mss (homeland) profile has been
added with all the logic required to retrieve image data and
metadata correctly.