Allow wand to fail if it doesn't have the right libs
If it can't do it, defer to pillow. This shouldn't really be
needed as if one can't do it then the other shouldn't be able
to as they probably use the same libs under the hood, but hey
this seems to work...

Also, wand is much faster at converting tifs than pillow, which
is why this code exists in the first place.