Use JSONResponse class throughout
There seems to be something up with the ORJSONResponse class when used
with the CORS middleware. If ORJSONResponse is used then the CORS headers
aren't included in the response but if JSONResponse is used then they are.
Pretty weird... Additional bonus is that we really don't need to be using
orjson for this project, speed is not a problem when generating JSON
responses, so binning the orjson stuff makes it simpler and leaves us with
one less dependency :)