Remove MSS collection record requirement
This commit removes the requirement for an MSS profile image to have
an associated specimen/index lot/artefact record in Elasticsearch.
This is because we rely on the APS part of the MSS to do this check
for us and because the IDs used for the images (specifically GUIDs)
are not guessable, so a user can't accidentally stumble upon images
they shouldn't be able to get to by incrementing the ID.

The side effects of this are reduced time processing requests because
we don't have to go ask Elasticsearch about the collections and
it gives us the ability to serve images not associated with collections
through this interface which is very useful!