Rewrite the way we get the source data stream to be more robust
It's better if we don't use an AsyncExitStack for this as it means we're leaving connections open after we're really done with them. For example, if we end up needing to go to the dams to get the data, we're leaving 2 connections open to the mss even though we're done with them.
Instead, we just open each url in turn and close the connection out when we're done with it.
Additionally, this commit adds exception handling so that we can track the URL we were attempting to access up the stack.