Move info.json generation out of profile and add extras
This solves a circular import issue which arose from generating the extra qualities in the ops module but then needing to import it into the profile base module. Additionally, it simplifies the info.json generation by removing the profile override ability (currently unused) and caching (was not well thought out as the part of the process that might take time, getting the info object, was not cached).