Proactively disable Pillow bomb errors in subprocesses
This seems to have become a problem post-loky, but only in production?
I would assume it's because Loky is doing something different in prod
when it spawns new processes and perhaps not completely copying the
existing memory state of the process? Who knows, either way this fixes
it and is inoffensive.