fix: remove image magick from jpg conversion code
Found a tiff which when converted just causes a seg fault from Image Magick due to an assertion error. Not sure why this happens so let's just remove Image Magick from the convert function and in turn just remove it completely from the project as this was the only place it was being used.
Shame as it was faster than pillow, hence why it was there, but you can't catch a seg fault and without knowing why the image is causing this problem, this is the best option right now.

Closes #41